// src/app/models/ref/joint-enum.ts
export enum JointEnum {
  RIGHT_KNEE = 'RIGHT_KNEE',
  LEFT_KNEE = 'LEFT_KNEE',
  LEFT_SHOULDER = 'LEFT_SHOULDER',
  RIGHT_SHOULDER = 'RIGHT_SHOULDER',
  LEFT_ELBOW = 'LEFT_ELBOW',
  RIGHT_ELBOW = 'RIGHT_ELBOW',
  CERVICAL = 'CERVICAL',
  LEFT_CERVICAL = 'LEFT_CERVICAL',
  RIGHT_CERVICAL = 'RIGHT_CERVICAL',
  HIP = 'HIP',
  LUMBAR = 'LUMBAR',
  LEFT_LUMBAR = 'LEFT_LUMBAR',
  RIGHT_LUMBAR = 'RIGHT_LUMBAR',
}